Extract page scrolling detection to custom hooks
Extracted for optional use on `useActiveChildScroll` hook
